+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3
 
+* Methods `execute` and `executeSafely` in `ErrorSandboxInterface` will return value of result of passed callback
 * Backward incompatible changes for ShutdownErrorException, FatalErrorException and SandboxException
 (setters have been removed, constructor and parent class have been changed)